ComponentDetails naming fixed - OrderDetails
